requirements.txt
playwright==1.49.1
pytest==8.4.2
google-genai==1.47.0
python-dotenv==1.0.1